Feature/Search feature implementation for booking

+++ server/controllers/futsalController.js
@@ -2,6 +2,7 @@ import futsalModel from "../model/Futsal.js"
 import { v4 as uuidv4 } from 'uuid';
 import multer from "multer";
 import bookingModel from "../model/Booking.js";
+import UserModel from "../model/User.js";
 
 
 const uniqueId = uuidv4();
@@ -30,12 +31,29 @@ const upload = multer({
   fileFilter : fileFilter,
 })
 
+const editStatus = async(req,res)=>{
+  try{
+    const {futsalId} = req.params;
+    const futsal = await futsalModel.findOne({_id:futsalId});
+    if(!futsal){
+      return res.status(404).json({message:"Futsal Not Found"});
+    }
+    const isOpen = futsal.isOpen;
+    const negation = !isOpen;
+    const updatedFutsal = await futsalModel.findByIdAndUpdate(futsalId, {isOpen: negation} , {
+      new: true, 
+    });
+    res.status(200).json({ message: 'Futsal updated successfully', futsal: updatedFutsal });
+  }catch(error){
+    console.log(error);
+    res.status(500).json({ message: "Error updating futsal" });
+  }
+}
 
 const createFutsal = async (req, res) => {
   try {
-    const { futsalName, futsalAddress, futsalDescription , addressLink} = req.body;
+    const { futsalName, futsalAddress, futsalDescription , addressLink, futsalContact} = req.body;
     const {userId} = req.params;
-    console.log(userId)
     const image = req.file.path;
 
     const existingFutsal = await futsalModel.findOne({
@@ -54,7 +72,9 @@ const createFutsal = async (req, res) => {
       address_link :`${addressLink}`,
       futsal_description: `${futsalDescription}`,
       vendorId : `${userId}`,
-      isValid : `${false}`
+      isValid : `${false}`,
+      isOpen: true,
+      futsal_contact:`${futsalContact}`
     });
     res.status(201).json({ message: "Futsal created successfully", futsal });
   } catch (error) {
@@ -97,10 +117,14 @@ const deleteFutsal = async (req, res) => {
     if (!futsalId) {
       return res.status(400).json({ msg: "No such futsal" });
     }
+    const futsal = await futsalModel.findOne({_id:futsalId})
+    if(!futsal){
+      return res.status(404).json({message:"Futsal Not Found"});
+    }
+    const userId = futsal.vendorId;
     const deletedFutsal = await futsalModel.deleteOne({ _id: futsalId });
     const deletedBookings = await bookingModel.deleteMany({ futsalId : futsalId});
-    console.log("DeletedBooking from futsal", deletedBookings)
-    console.log("Deleted Futsals", deletedFutsal);
+    const deletedVendor = await UserModel.deleteOne({_id: userId}) 
     return res.status(201).json({ message: "Futsal deleted", deleteFutsal });
   } catch (error) {
     console.log(error);
@@ -111,11 +135,10 @@ const deleteFutsal = async (req, res) => {
 const editFutsal = async (req, res) => {
   try {
     const {futsalId} = req.params; 
-  const updatedData = req.body;  
+    const updatedData = req.body;  
 
     const updatedFutsal = await futsalModel.findByIdAndUpdate(futsalId, updatedData, {
       new: true, 
-      runValidators: true  // Validation  yet to be done
     });
 
     if (!updatedFutsal) {
@@ -128,6 +151,19 @@ const editFutsal = async (req, res) => {
   }
 };
 
+const getFutsalById = async (req,res) =>{
+  try{
+    const {futsalId} = req.params;
+    const futsal = await futsalModel.findOne({_id:futsalId});
+    if(!futsal){
+      return res.status(401),json({message: "Futsal not found"});
+    }
+    res.status(200).json({message:"Futsal retrieved sucessfully", futsal})
+  }catch(err){
+    console.log(err);
+    return res.status(500).json({message: "INternal server error"});
+  }
 
+}
 
-export { getAllFutsals, createFutsal, getVendorSpecificFutsal , editFutsal, upload, deleteFutsal};
+export { getAllFutsals, createFutsal, getVendorSpecificFutsal , editFutsal, upload, deleteFutsal, getFutsalById, editStatus};
